+++ config/forex.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'email_recipients' => env('FOREX_EMAIL_RECIPIENTS', ''),
+];